Feature connect sign up

+++ Emby.Mobile.ViewModels/ViewModels/EmbyConnectViewModel.cs
@@ -1,7 +1,9 @@
 ﻿using System;
+using System.Threading.Tasks;
 using Emby.Mobile.Core.Interfaces;
 using Emby.Mobile.Helpers;
 using GalaSoft.MvvmLight.Command;
+using JetBrains.Annotations;
 using MediaBrowser.Model.ApiClient;
 using MediaBrowser.Model.Net;
 
@@ -9,7 +11,7 @@ namespace Emby.Mobile.ViewModels
 {
     public class EmbyConnectViewModel : ViewModelBase
     {
-        public EmbyConnectViewModel(IServices services) 
+        public EmbyConnectViewModel(IServices services)
             : base(services)
         {
         }
@@ -27,41 +29,60 @@ namespace Emby.Mobile.ViewModels
             {
                 return new RelayCommand(async () =>
                 {
-                    try
-                    {
-                        SetProgressBar(GetLocalizedString("SysTraySigningIn"));
-
-                        var success = await AuthenticationService.LoginWithConnect(Username, Password);
-
-                        if (success)
-                        {
-                            var result = await Services.ConnectionManager.Connect();
-                            if (result.State == ConnectionState.SignedIn && result.Servers.Count == 1)
-                            {
-                                Services.ServerInfo.SetServerInfo(result.Servers[0]);
-                                Services.ServerInfo.Save();
-                            }
-
-                            AuthenticationService.SetConnectUser(result.ConnectUser);
-
-                            await ConnectHelper.HandleConnectState(result, Services, ApiClient);
-
-                            Services.NavigationService.RemoveBackEntry();
-                        }
-                    }
-                    catch (HttpException hex)
-                    {
-
-                    }
-                    catch (Exception ex)
-                    {
-
-                    }
-                    finally
-                    {
-                        SetProgressBar();
-                    }
+                    await SignIn();
                 });
+            }
+        }
+
+        public RelayCommand SignUpCommand
+        {
+            get
+            {
+                return new RelayCommand(() =>
+                {
+                    Services.NavigationService.NavigateToEmbyConnectSignUp();
+                });
+            }
+        }
+
+        private async Task SignIn()
+        {
+            try
+            {
+                if (!CanSignIn)
+                {
+                    return;
+                }
+
+                SetProgressBar(GetLocalizedString("SysTraySigningIn"));
+
+                var success = await AuthenticationService.LoginWithConnect(Username, Password);
+
+                if (success)
+                {
+                    var result = await Services.ConnectionManager.Connect();
+                    if (result.State == ConnectionState.SignedIn && result.Servers.Count == 1)
+                    {
+                        Services.ServerInfo.SetServerInfo(result.Servers[0]);
+                        Services.ServerInfo.Save();
+                    }
+
+                    AuthenticationService.SetConnectUser(result.ConnectUser);
+
+                    await ConnectHelper.HandleConnectState(result, Services, ApiClient);
+
+                    Services.NavigationService.RemoveBackEntry();
+                }
+            }
+            catch (HttpException hex)
+            {
+            }
+            catch (Exception ex)
+            {
+            }
+            finally
+            {
+                SetProgressBar();
             }
         }
